requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
httptools==0.3.0
idna==3.2
motor==2.3.1
multidict==5.1.0
odmantic==0.3.5
Pillow==8.3.1
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pymongo==3.12.0
pyOpenSSL==20.0.1
sanic==21.6.2
sanic-routing==0.7.1
six==1.16.0
typing-extensions==3.10.0.0
websockets==9.1
yarl==1.6.3
